requirements.txt
agents-arcade>=0.0.4
arcadepy>=1.0.0  
python-dotenv>=1.0.0
slack-sdk>=3.31.0